Create websitescrape
This is my first attempt at scraping a website, while using Request-HTML Library. This scraper is able to pull the data of the headline, summary and youtube links from a demonstration website. I feel this is a good start before translating the code to work on the RNP website. This project required a lot of parsing, as sometimes website source code does not have the information you want in the most accessible way. I have now increased my confidence in inspecting HTML.

It was recommended to use Requests-HTML Library because it is god for referencing links better than Beautiful Soup. In addition, it has the ability to grab text from Java Script, it gave example code on how asynchronous parsing saves time, as well as, it can spoof a user agent to avoid being blocked for hitting a website too fast.